Add files via upload
EmailListDiff.py is a Python program that prompts user to enter two files. First one must be a .csv file with additional email list and the second file must be a .csv file with master list. The program outputs a diff.csv file with emails that are not present in the master list.